KEY-10 Split the codebase into modules
