+++ src/index.ts
@@ -1,0 +1,3 @@
+import { start } from "./app";
+
+start(window);
